# Object3D performance/memory improvements