My project requires the creation of large numbers of Object3D instances, so I've made some speed and memory usage improvements for that scenario. This first submission is in two parts, both dealing with reducing closures within the Object3D constructor.

The first commit is simply moving the default onBeforeRender/onAfterRender empty callback functions to the prototype. Surprisingly this also increases their call speed per frame as well, presumably since it allows the JS engine to optimize away the empty functions better.

The second commit changes the quaternion/rotation synchronization to utilize member cross-references and shared onChangeCallback functions rather than two closures per pair. This relies on the onChangeCallback being executed with the proper `this` but that appears to always be the case.

Combined measurements, using 100K Object3D instances:
- constructor exec time: 1620ms -> 1200ms
- onBeforeRender/onAfterRender call time: 4.5ms -> 1.1ms
- retained size: 133MB -> 101MB