@mrdoob 2017-05-05
The onBeforeRender/onAfterRender change is a no brainer. The euler/quaternion one needs some more thinking/clean up though. Do you mind splitting the PR in two?