@lojjic 2017-05-05
Not a problem; I kept them as separate commits with that in mind. Expect two new PRs shortly.